requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coverage==6.3
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
numpy==1.22.1
packaging==21.3
pandas==1.3.0
pdf2image==1.16.0
Pillow==8.3.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0.post1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
